+++ Tests/KeyboardKitSwiftUITests/System/SystemKeyboardButtonRowTests.swift
@@ -12,19 +12,19 @@ import KeyboardKit
 import SwiftUI
 @testable import KeyboardKitSwiftUI
 
-class SystemKeyboardBottomRowTests: QuickSpec {
+class SystemKeyboardButtonRowTests: QuickSpec {
 
     override func spec() {
         
-        describe("system keyboard bottom row") {
+        describe("system keyboard button row") {
             
             it("can be created with the default button builder") {
-                let button = SystemKeyboardBottomRow(leftmostAction: .command)
+                let button = SystemKeyboardButtonRow(actions: [.command])
                 expect(button).toNot(beNil())
             }
             
             it("can be created with a custom button builder") {
-                let button = SystemKeyboardBottomRow(leftmostAction: .command, buttonBuilder: { _ in AnyView(Text("HEJ")) })
+                let button = SystemKeyboardButtonRow(actions: [.command], buttonBuilder: { _ in AnyView(Text("HEJ")) })
                 expect(button).toNot(beNil())
             }
         }
@@ -32,61 +32,9 @@ class SystemKeyboardBottomRowTests: QuickSpec {
         describe("standard button builder") {
             
             it("returns a view") {
-                let builder = SystemKeyboardBottomRow.standardButtonBuilder()
+                let builder = SystemKeyboardButtonRow.standardButtonBuilder
                 let result = builder(.backspace)
                 expect(result).toNot(beNil())
-            }
-        }
-        
-        describe("action collection") {
-            
-            it("is correctly setup when context needs input mode switch key") {
-                let context = MockKeyboardContext()
-                context.needsInputModeSwitchKey = true
-                let row = SystemKeyboardBottomRow(leftmostAction: .control) { _ in fatalError() }
-                let actions = row.actions(for: context)
-                expect(actions).to(equal([
-                    .control,
-                    .nextKeyboard,
-                    .space,
-                    .keyboardType(.emojis),
-                    .newLine
-                ]))
-            }
-            
-            it("is correctly setup when context doesn't need input mode switch key") {
-                let context = MockKeyboardContext()
-                context.needsInputModeSwitchKey = false
-                let row = SystemKeyboardBottomRow(leftmostAction: .escape) { _ in fatalError() }
-                let actions = row.actions(for: context)
-                expect(actions).to(equal([
-                    .escape,
-                    .keyboardType(.emojis),
-                    .space,
-                    .keyboardType(.images),
-                    .newLine
-                ]))
-            }
-        }
-        
-        describe("view collection") {
-            
-            it("is correctly setup when context needs input mode switch key") {
-                let context = MockKeyboardContext()
-                context.needsInputModeSwitchKey = true
-                let row = SystemKeyboardBottomRow(leftmostAction: .control) { _ in fatalError() }
-                let actions = row.actions(for: context)
-                let views = row.actions(for: context)
-                expect(views.count).to(equal(actions.count))
-            }
-            
-            it("is correctly setup when context doesn't need input mode switch key") {
-                let context = MockKeyboardContext()
-                context.needsInputModeSwitchKey = false
-                let row = SystemKeyboardBottomRow(leftmostAction: .escape) { _ in fatalError() }
-                let actions = row.actions(for: context)
-                let views = row.actions(for: context)
-                expect(views.count).to(equal(actions.count))
             }
         }
     }